Disable Background Modes in the Examples app
Since the Estimote SDK is built on top of Core Location, Background
Modes are not needed for beacons to work in the background--the "always"
authorization to access Location Services is all that's needed.

Having Background Modes enabled when your app does not in fact need them
(as is the case with iBeacon) will result in the app getting rejected by
the App Store review team!